+++ src/Board320_240_menu.cpp
@@ -30,6 +30,27 @@ String Board320_240::menuItemText(int16_t menuItemId, String title)
     sprintf(tmpStr1, "[%s]", APP_VERSION);
     suffix = tmpStr1;
     break;
+  case MENU_WIFI:
+    if (liveData->settings.wifiEnabled != 1 || WiFi.status() != WL_CONNECTED)
+    {
+      suffix = "[not_connected]";
+    }
+    else if (wifiTransferredBytes >= (1024UL * 1024UL))
+    {
+      sprintf(tmpStr1, "[%luMB]", (unsigned long)(wifiTransferredBytes / (1024UL * 1024UL)));
+      suffix = tmpStr1;
+    }
+    else if (wifiTransferredBytes >= 1024UL)
+    {
+      sprintf(tmpStr1, "[%lukB]", (unsigned long)(wifiTransferredBytes / 1024UL));
+      suffix = tmpStr1;
+    }
+    else
+    {
+      sprintf(tmpStr1, "[%luB]", (unsigned long)wifiTransferredBytes);
+      suffix = tmpStr1;
+    }
+    break;
   // TODO: Why is do these cases not match the vehicle type id?
   case VEHICLE_TYPE_IONIQ_2018_28:
     prefix = (liveData->settings.carType == CAR_HYUNDAI_IONIQ_2018) ? ">" : "";
@@ -198,7 +219,7 @@ String Board320_240::menuItemText(int16_t menuItemId, String title)
     sprintf(tmpStr1, "%s", liveData->settings.mqttPubTopic);
     suffix = tmpStr1;
     break;
-  case MENU_REMOTE_UPLOAD_CONTRIBUTE_ANONYMOUS_DATA_TO_EVDASH_DEV_TEAM:
+  case MENU_REMOTE_UPLOAD_CONTRIBUTE_DATA_TO_EVDASH_DEV_TEAM:
     suffix = (liveData->settings.contributeData == 0) ? "[off]" : "[on]";
     break;
   case MENU_SDCARD:
@@ -1165,13 +1186,18 @@ void Board320_240::menuItemClick()
       return;
     }
     break;
-    case MENU_REMOTE_UPLOAD_CONTRIBUTE_ANONYMOUS_DATA_TO_EVDASH_DEV_TEAM:
+    case MENU_REMOTE_UPLOAD_CONTRIBUTE_DATA_TO_EVDASH_DEV_TEAM:
       liveData->settings.contributeData = (liveData->settings.contributeData == 1) ? 0 : 1;
       if (liveData->settings.contributeData == 0)
       {
         liveData->params.contributeStatus = CONTRIBUTE_NONE;
         contributeStatusSinceMs = 0;
+        nextContributeCycleAtMs = 0;
         liveData->clearContributeRawFrames();
+      }
+      else
+      {
+        nextContributeCycleAtMs = millis();
       }
       showMenu();
       return;
@@ -1186,6 +1212,7 @@ void Board320_240::menuItemClick()
       liveData->params.lastContributeSent = liveData->params.currentTime;
       liveData->params.contributeStatus = CONTRIBUTE_WAITING;
       contributeStatusSinceMs = millis();
+      nextContributeCycleAtMs = millis() + kContributeCycleIntervalMs;
       if (liveData->settings.contributeJsonType == CONTRIBUTE_JSON_TYPE_V2 &&
           liveData->settings.commType == COMM_TYPE_CAN_COMMU &&
           commInterface != nullptr)
@@ -1563,11 +1590,7 @@ void Board320_240::menuItemClick()
       break;
     // Version
     case MENU_APP_VERSION:
-      if (confirmMessage("Confirm action", "Do you want to run OTA?"))
-      {
-        showMenu();
-        otaUpdate();
-      }
+      displayMessage("App version", APP_VERSION);
       showMenu();
       return;
       break;